tests/agents/requirements.txt
agno
python-dotenv
ollama
uv